services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/core/docker_utils.py
import logging
from contextlib import asynccontextmanager
from typing import Any, AsyncGenerator, Dict

import aiodocker
from aiodocker.utils import clean_filters

from .errors import UnexpectedDockerError, VolumeNotFoundError

logger = logging.getLogger(__name__)


@asynccontextmanager
async def docker_client() -> AsyncGenerator[aiodocker.Docker, None]:
    docker = aiodocker.Docker()
    try:
        yield docker
    except aiodocker.exceptions.DockerError as error:
        logger.debug("An unexpected Docker error occurred", stack_info=True)
        raise UnexpectedDockerError(
            message=error.message, status=error.status
        ) from error
    finally:
        await docker.close()


async def get_volume_by_label(label: str) -> Dict[str, Any]:
    async with docker_client() as docker:
        filters = {"label": [f"source={label}"]}
        params = {"filters": clean_filters(filters)}
        data = await docker._query_json(  # pylint: disable=protected-access
            "volumes", method="GET", params=params
        )
        volumes = data["Volumes"]
        logger.debug(  # pylint: disable=logging-fstring-interpolation
            f"volumes query for {label=} {volumes=}"
        )
        if len(volumes) != 1:
            raise VolumeNotFoundError(volumes)
        volume_details = volumes[0]
        return volume_details  # type: ignore
